Continued the models refactor by creating a file for the base models
