+++ dna/__main__.py
@@ -8,7 +8,8 @@ import typing
 import uuid
 
 from dna.data import get_data, preprocess_data, split_data, group_json_objects
-from dna.models.models import get_model, ModelBase
+from dna.models.models import get_model
+from dna.models.base_models import ModelBase
 from dna.problems import get_problem, ProblemBase
