+++ dna/models/__init__.py
@@ -1,0 +1,2 @@
+class ModelNotFitError(Exception):
+    pass
